requirements.txt
numpy>=1.23.0
pandas==1.4.2
Pillow==9.4.0
pydeck==0.8.0b4
st_pages==0.3.1
streamlit==1.17.0
streamlit_lottie==0.0.3
